# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clare Pastries API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  slug: string;
  name: string;
  description: string;
  priceKes: number;
  category: string;
  imageUrl: string;
  featured: boolean;
  inStock: boolean;
  servings?: string;
}

export type CustomOrderInputFulfillment =
  (typeof CustomOrderInputFulfillment)[keyof typeof CustomOrderInputFulfillment];

export const CustomOrderInputFulfillment = {
  delivery: "delivery",
  pickup: "pickup",
} as const;

export interface CustomOrderInput {
  fullName: string;
  phone: string;
  email?: string;
  occasion: string;
  description: string;
  servings?: number;
  preferredDate?: string;
  budget?: string;
  fulfillment: CustomOrderInputFulfillment;
}

export interface CustomOrder {
  id: string;
  fullName: string;
  phone: string;
  email?: string;
  occasion: string;
  description: string;
  servings?: number;
  preferredDate?: string;
  budget?: string;
  fulfillment: string;
  createdAt: string;
}

export interface ContactInput {
  name: string;
  email?: string;
  phone?: string;
  message: string;
}

export interface ContactMessage {
  id: string;
  name: string;
  email?: string;
  phone?: string;
  message: string;
  createdAt: string;
}

export interface Review {
  id: string;
  author: string;
  rating: number;
  body: string;
  location?: string;
  createdAt: string;
}

export interface GalleryItem {
  id: string;
  title: string;
  category: string;
  imageUrl: string;
  productId?: string;
}

export interface Stats {
  totalProducts: number;
  featuredCount: number;
  avgRating?: number;
  deliveryMins: string;
  cityServed: string;
}

export type ListProductsParams = {
  category?: ListProductsCategory;
  featured?: boolean;
};

export type ListProductsCategory =
  (typeof ListProductsCategory)[keyof typeof ListProductsCategory];

export const ListProductsCategory = {
  all: "all",
  cakes: "cakes",
  pastries: "pastries",
  breads: "breads",
  seasonal: "seasonal",
} as const;

export type ListGalleryParams = {
  category?: string;
};