# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Clare Pastries API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [all, cakes, pastries, breads, seasonal]
        - name: featured
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        priceKes: { type: number }
        category: { type: string }
        imageUrl: { type: string }
        featured: { type: boolean }
        inStock: { type: boolean }
        servings: { type: string }
      required: [id, slug, name, description, priceKes, category, imageUrl, featured, inStock]
...