# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Clare Pastries API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  slug: string;
  name: string;
  description: string;
  priceKes: number;
  category: string;
  imageUrl: string;
  featured: boolean;
  inStock: boolean;
  servings?: string;
}